requirements.txt
fastapi
uvicorn
tensorflow==2.15.0
pydantic
numpy
Pillow